Move querykey shape declaration into type folder
