+++ src/app/_hooks/use-supplier-tiles.ts
@@ -2,7 +2,7 @@
 
 import { useQuery, useQueryClient } from '@tanstack/react-query'
 
-import { tileKeys } from '@/app/_hooks/queryKeys'
+import { tileKeys } from '@/app/_types/queryKeys'
 import { SupplierTilesGetRequestParams, SupplierTilesGetResponseBody } from '@/app/api/suppliers/[id]/tiles/route'
 import { buildQueryParams } from '@/utils/api-helpers'
 import { DEFAULT_STALE_TIME } from '@/utils/constants'
