+++ src/app/t/[tileId]/page.tsx
@@ -3,7 +3,7 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { notFound } from 'next/navigation'
 
-import { tileKeys } from '@/app/_hooks/queryKeys'
+import { tileKeys } from '@/app/_types/queryKeys'
 import { SaveTileButton } from '@/components/tiles/save-button'
 import { Area } from '@/components/ui/area'
 import { Section } from '@/components/ui/section'
